Backend/pyproject.toml
[project]
name = "ai-job-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apify-client>=2.4.1",
    "google-generativeai>=0.8.6",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
]
